refactor: extract node model
