+++ lua/bookmarks/domain/init.lua
@@ -1,6 +1,7 @@
 local location = require("bookmarks.domain.location")
 local bookmark = require("bookmarks.domain.bookmark")
 local bookmark_list = require("bookmarks.domain.bookmark_list")
+local bookmark_node = require("bookmarks.domain.bookmark_node")
 local project = require("bookmarks.domain.project")
 local _type = require("bookmarks.domain.type")
 
@@ -16,6 +17,7 @@ local BookmarkModule = {
   location = location,
   bookmark = bookmark,
   bookmark_list = bookmark_list,
+  node = bookmark_node,
   project = project,
   type = _type,
 }
